# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dental clinic API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Service {
  id: number;
  nameAr: string;
  nameEn: string;
  descriptionAr: string;
  descriptionEn: string;
  /** Duration in minutes */
  duration: number;
  icon: string;
}

export interface Doctor {
  id: number;
  nameAr: string;
  nameEn: string;
  specialtyAr: string;
  specialtyEn: string;
  bio: string;
  /** Years of experience */
  experience: number;
  /** @nullable */
  imageUrl: string | null;
}

export type AppointmentStatus = typeof AppointmentStatus[keyof typeof AppointmentStatus];


export const AppointmentStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Appointment {
  id: number;
  patientName: string;
  patientPhone: string;
  /** @nullable */
  patientEmail?: string | null;
  serviceId: number;
  doctorId: number;
  date: string;
  time: string;
  /** @nullable */
  notes?: string | null;
  status: AppointmentStatus;
  createdAt: string;
}

export interface AppointmentInput {
  patientName: string;
  patientPhone: string;
  patientEmail?: string;
  serviceId: number;
  doctorId: number;
  date: string;
  time: string;
  notes?: string;
}

export type AppointmentUpdateStatus = typeof AppointmentUpdateStatus[keyof typeof AppointmentUpdateStatus];


export const AppointmentUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface AppointmentUpdate {
  status?: AppointmentUpdateStatus;
  date?: string;
  time?: string;
  notes?: string;
}

export interface TimeSlot {
  time: string;
  available: boolean;
}

export interface ClinicStats {
  totalAppointments: number;
  todayAppointments: number;
  pendingAppointments: number;
  completedAppointments: number;
}

export type GetAvailableSlotsParams = {
date: string;
doctorId?: number;
};