# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Dental clinic API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: appointments
    description: Appointment booking
  - name: services
    description: Dental services
  - name: doctors
    description: Clinic doctors
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all dental services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"

  /doctors:
    get:
      operationId: listDoctors
      tags: [doctors]
      summary: List all clinic doctors
      responses:
        "200":
          description: List of doctors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doctor"

  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book a new appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Validation error

  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Cancel an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /clinic/stats:
    get:
      operationId: getClinicStats
      tags: [appointments]
      summary: Get clinic dashboard stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicStats"

  /appointments/available-slots:
    get:
      operationId: getAvailableSlots
      tags: [appointments]
      summary: Get available time slots for a date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
        - name: doctorId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Available time slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSlot"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Service:
      type: object
      required: [id, nameAr, nameEn, descriptionAr, descriptionEn, duration, icon]
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        descriptionAr:
          type: string
        descriptionEn:
          type: string
        duration:
          type: integer
          description: Duration in minutes
        icon:
          type: string

    Doctor:
      type: object
      required: [id, nameAr, nameEn, specialtyAr, specialtyEn, bio, experience, imageUrl]
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        specialtyAr:
          type: string
        specialtyEn:
          type: string
        bio:
          type: string
        experience:
          type: integer
          description: Years of experience
        imageUrl:
          type: ["string", "null"]

    Appointment:
      type: object
      required: [id, patientName, patientPhone, serviceId, doctorId, date, time, status, createdAt]
      properties:
        id:
          type: integer
        patientName:
          type: string
        patientPhone:
          type: string
        patientEmail:
          type: ["string", "null"]
        serviceId:
          type: integer
        doctorId:
          type: integer
        date:
          type: string
        time:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        createdAt:
          type: string

    AppointmentInput:
      type: object
      required: [patientName, patientPhone, serviceId, doctorId, date, time]
      properties:
        patientName:
          type: string
        patientPhone:
          type: string
        patientEmail:
          type: string
        serviceId:
          type: integer
        doctorId:
          type: integer
        date:
          type: string
        time:
          type: string
        notes:
          type: string

    AppointmentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        date:
          type: string
        time:
          type: string
        notes:
          type: string

    TimeSlot:
      type: object
      required: [time, available]
      properties:
        time:
          type: string
        available:
          type: boolean

    ClinicStats:
      type: object
      required: [totalAppointments, todayAppointments, pendingAppointments, completedAppointments]
      properties:
        totalAppointments:
          type: integer
        todayAppointments:
          type: integer
        pendingAppointments:
          type: integer
        completedAppointments:
          type: integer