# lib/api-zod/src/generated/types/appointment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dental clinic API
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentStatus } from './appointmentStatus';

export interface Appointment {
  id: number;
  patientName: string;
  patientPhone: string;
  /** @nullable */
  patientEmail?: string | null;
  serviceId: number;
  doctorId: number;
  date: string;
  time: string;
  /** @nullable */
  notes?: string | null;
  status: AppointmentStatus;
  createdAt: string;
}