# lib/api-zod/src/generated/types/appointmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dental clinic API
 * OpenAPI spec version: 0.1.0
 */

export interface AppointmentInput {
  patientName: string;
  patientPhone: string;
  patientEmail?: string;
  serviceId: number;
  doctorId: number;
  date: string;
  time: string;
  notes?: string;
}