# lib/api-zod/src/generated/types/appointmentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dental clinic API
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentUpdateStatus } from './appointmentUpdateStatus';

export interface AppointmentUpdate {
  status?: AppointmentUpdateStatus;
  date?: string;
  time?: string;
  notes?: string;
}