# lib/api-zod/src/generated/types/doctor.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dental clinic API
 * OpenAPI spec version: 0.1.0
 */

export interface Doctor {
  id: number;
  nameAr: string;
  nameEn: string;
  specialtyAr: string;
  specialtyEn: string;
  bio: string;
  /** Years of experience */
  experience: number;
  /** @nullable */
  imageUrl: string | null;
}