# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Dental clinic API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  nameAr: string;
  nameEn: string;
  descriptionAr: string;
  descriptionEn: string;
  /** Duration in minutes */
  duration: number;
  icon: string;
}