# lib/db/src/schema/doctors.ts
import { pgTable, text, serial, integer } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const doctorsTable = pgTable("doctors", {
  id: serial("id").primaryKey(),
  nameAr: text("name_ar").notNull(),
  nameEn: text("name_en").notNull(),
  specialtyAr: text("specialty_ar").notNull(),
  specialtyEn: text("specialty_en").notNull(),
  bio: text("bio").notNull(),
  experience: integer("experience").notNull(),
  imageUrl: text("image_url"),
});

export const insertDoctorSchema = createInsertSchema(doctorsTable).omit({ id: true });
export type InsertDoctor = z.infer<typeof insertDoctorSchema>;
export type Doctor = typeof doctorsTable.$inferSelect;